requirement.txt
boto3==1.4.4
botocore==1.5.27
Django==1.10.6
django-localflavor==1.4.1
djangorestframework==3.6.2
docutils==0.13.1
jmespath==0.9.2
mysqlclient==1.3.10
pkg-resources==0.0.0
python-dateutil==2.6.0
requests==2.13.0
s3transfer==0.1.10
six==1.10.0
